added few things in Javascripts
I added a few things in Javascripts the evening of yestarday trying to figure out what I can do in order to start on my logic for Javascript, it did not go as I planned and I am submitting to GitHub what I have. Also in order to use Netlify, I need to change my files names to the conventional "index.html" as well ass css to "style.css".